// src/shared/enums/routers.enum.ts
export enum Routers {
  dashboardDetail = 'DashboardDetail',
  chartBuilder = 'ChartBuilder',
  dataBuilder = 'DataBuilder',
  notFound = 'NotFound',

  login = 'Login',
  signup = 'Signup',
  forgotPassword = 'ForgotPassword',
  passwordRecovery = 'PasswordRecovery',
  resendEmail = 'ResendEmail',
  directVerify = 'DirectVerify',

  subDirectory = 'SubDirectory',
  mydata = 'MyData',

  trackingProfile = 'UserProfile',
  trackingProfileDetail = 'UserProfileDetails',
  userManagement = 'UserManagement',
  userManagementDetails = 'UserManagementDetails',
  baseRoute = '',

  sharedWithMe = 'Shared',
  subShared = 'SubShared',
  directoryShare = 'DirectoryShare',

  dataManagement = 'DataManagement',
  dataSchema = 'DataSchema',
  queryEditor = 'QueryEditor',
  dataRelationship = 'DataRelationship',

  dataIngestion = 'DataIngestion',
  dataSource = 'dataSource',
  job = 'job',
  jobHistory = 'jobHistory'
}